+++ backend/src/hatchling/builders/hooks/plugin/hooks.py
@@ -1,7 +1,8 @@
 from ....plugin import hookimpl
 from ..custom import CustomBuildHook
+from ..version import VersionBuildHook
 
 
 @hookimpl
 def hatch_register_build_hook():
-    return CustomBuildHook
+    return [CustomBuildHook, VersionBuildHook]