+++ docs/history.md
@@ -74,11 +74,13 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 
 ***Added:***
 
+- Add `version` build hook
 - Support PEP 561
 
 ***Fixed:***
 
 - Don't write empty entry points file for `wheel` targets if there are no entry points defined
+- Allow metadata hooks to set the `version` in all cases
 
 ### [1.3.1](https://github.com/pypa/hatch/releases/tag/hatchling-v1.3.1) - 2022-05-30 ### {: #hatchling-v1.3.1 }